[!9][WHISPER] Script for batched ASR with Whisper
To enable the transcription of large datasests with Whisper, we need an easy-to-use and easy-to-customize tool that performs batching.

Adds a script that performs transcription with Whisper and supports configurable `AudioIterator`s that specify the list of audios to transcribe. The current MR contains the `AudioIterator` for Voxpopuli but this has to be extended in the future.

UTs and manual execution. Sample usage on local machine:

```
whisper-gen --hf-model-name openai/whisper-large-v3 --audio-iterator voxpopuli --audio-iterator-config config.yaml -bs 2 -bm 5 -o ooo.tsv
```