[!12][WHISPER] Read with soundfile instead of ffmpeg
# Why is the change needed?

The huggingface library uses ffmpeg to read audio files. THis requires the creation of a new process to read every audio, causing a significant overhead for the OS.

# What changes does the patch introduce?

Use soundfile to read the audios inside the python process instead of relying on ffmpeg so that we do not spawn a new process to read each audio file.

# How was this patch tested?

manual execution and comparison with previous outputs