[!13][WHISPER] Support restart Whisper generation from the last generated output
# Why is the change needed?

When transcribing huge datasets (like the Voxpopuli unlabelled data), jobs may take a lot of time to complete and in many clusters we have time limits after which the job is killed regardless of whether it finished or not. In this situation, it is useful to avoid restarting the transcription from scratch, but, instead, perform the transcription only of the missing elements and append them to the output.

# What changes does the patch introduce?

Adds an option to append to an existing output file the missing samples rather than starting from scratch and ovverriding it.

# How was this patch tested?

added UTs