[!14][WHISPER][BUG] Fix processing of elements with weidly-formatted id
# Why is the change needed?

I found that some rows (in Estonian) have weird IDs like this one `20090224-1736-committee-afet_20090224-1736-COMMITTEE-AFET_et` which do not conform to the usual format, but are still valid samples and the current code ignores them.

# What changes does the patch introduce?

Fix the parsing of the `event_id` field so that it supports also these samples.

# How was this patch tested?

added UT